rename files to snake_case format
