+++ cpp/include/ray/api/ray_api.h
@@ -1,14 +1,14 @@
 
 #pragma once
 
-#include <ray/api/Blob.h>
+#include <ray/api/blob.h>
 #include <cstdint>
 #include <memory>
 #include <typeinfo>
 #include <vector>
 
-#include "ray/api/UniqueId.h"
-#include "ray/util/type-util.h"
+#include "ray/api/uniqueId.h"
+#include "ray/util/type_util.h"
 
 namespace ray {
