+++ cpp/include/ray/api/ray_mode.h
