+++ cpp/include/ray/api/ray_object.h
@@ -4,7 +4,7 @@
 #include <memory>
 #include <utility>
 
-#include "UniqueId.h"
+#include "uniqueId.h"
 
 namespace ray {
