+++ cpp/include/ray/util/type_util.h
