+++ cpp/src/ray/api/UniqueId.cc
@@ -1,5 +1,5 @@
 
-#include <ray/api/UniqueId.h>
+#include <ray/api/uniqueId.h>
 #include <random>
 
 namespace ray {
