+++ cpp/src/ray/runtime/ray_native_runtime.h
@@ -1,7 +1,7 @@
 
 #pragma once
 
-#include "RayRuntime.h"
+#include "ray_runtime.h"
 
 namespace ray {
