+++ cpp/src/ray/spi/Worker.cc
@@ -1,5 +1,5 @@
 
-#include "Worker.h"
+#include "worker.h"
 
 namespace ray {
