+++ cpp/src/ray/spi/object_proxy.cc
@@ -1,5 +1,5 @@
 
-#include "ObjectProxy.h"
+#include "object_proxy.h"
 
 #include <memory>
 #include <utility>
