+++ cpp/src/ray/spi/object_proxy.h
@@ -3,11 +3,11 @@
 
 #include <memory>
 
-#include <ray/api/Blob.h>
-#include <ray/api/UniqueId.h>
-#include <ray/util/type-util.h>
+#include <ray/api/blob.h>
+#include <ray/api/uniqueId.h>
+#include <ray/util/type_util.h>
 
-#include "ObjectInterface.h"
+#include "object_interface.h"
 
 namespace ray {
