+++ cpp/src/ray/spi/task_spec.cc
@@ -1,4 +1,4 @@
-#include "TaskSpec.h"
+#include "task_spec.h"
 
 namespace ray {
