+++ cpp/src/ray/spi/task_spec.h
@@ -4,8 +4,8 @@
 #include <list>
 #include <vector>
 
-#include <ray/api/Blob.h>
-#include <ray/api/UniqueId.h>
+#include <ray/api/blob.h>
+#include <ray/api/uniqueId.h>
 
 namespace ray {
