+++ cpp/src/ray/test/api_test.cc
